requirements.txt
# Core ML and Deep Learning
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0

# Data processing and analysis
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
imageio>=2.19.0
pillow>=8.3.0

# Progress and utilities
tqdm>=4.62.0

# Experiment tracking
wandb>=0.13.0

# Optional: Advanced models (if needed)
transformers>=4.20.0

# Development and debugging
ipython>=7.25.0
jupyter>=1.0.0